+++ modules/browser-node/src/services/store.spec.ts
@@ -46,13 +46,7 @@ describe("store", () => {
   describe("getChannelStateByParticipants", () => {
     it("should work (regardless of order)", async () => {
       const channel = createTestChannelState("deposit");
-      await store.saveChannelState(channel, {
-        channelFactoryAddress: channel.networkContext.channelFactoryAddress,
-        chainId: channel.networkContext.chainId,
-        aliceSignature: channel.latestUpdate.aliceSignature,
-        bobSignature: channel.latestUpdate.bobSignature,
-        state: channel,
-      });
+      await store.saveChannelState(channel);
 
       expect(
         await store.getChannelStateByParticipants(
@@ -79,17 +73,7 @@ describe("store", () => {
         channelAddress: channel.channelAddress,
         meta: { routingId: getRandomBytes32() },
       });
-      await store.saveChannelState(
-        channel,
-        {
-          channelFactoryAddress: channel.networkContext.channelFactoryAddress,
-          chainId: channel.networkContext.chainId,
-          aliceSignature: channel.latestUpdate.aliceSignature,
-          bobSignature: channel.latestUpdate.bobSignature,
-          state: channel,
-        },
-        transfer,
-      );
+      await store.saveChannelState(channel, transfer);
 
       expect(await store.getTransferByRoutingId(channel.channelAddress, transfer.meta!.routingId)).to.be.deep.eq(
         transfer,
@@ -103,13 +87,7 @@ describe("store", () => {
       const c2 = createTestChannelState("deposit", { channelAddress: mkAddress("0xccc2222") });
       await Promise.all(
         [c1, c2].map(c => {
-          return store.saveChannelState(c, {
-            channelFactoryAddress: c.networkContext.channelFactoryAddress,
-            chainId: c.networkContext.chainId,
-            aliceSignature: c.latestUpdate.aliceSignature,
-            bobSignature: c.latestUpdate.bobSignature,
-            state: c,
-          });
+          return store.saveChannelState(c);
         }),
       );
       const retrieved = await store.getChannelStates();
@@ -121,13 +99,7 @@ describe("store", () => {
   it("should save transaction responses and receipts", async () => {
     // Load store with channel
     const setupState = createTestChannelState("setup");
-    await store.saveChannelState(setupState, {
-      channelFactoryAddress: setupState.networkContext.channelFactoryAddress,
-      chainId: setupState.networkContext.chainId,
-      aliceSignature: setupState.latestUpdate.aliceSignature,
-      bobSignature: setupState.latestUpdate.bobSignature,
-      state: setupState,
-    });
+    await store.saveChannelState(setupState);
 
     const response = createTestTxResponse();
 
@@ -190,13 +162,7 @@ describe("store", () => {
 
   it("should save and retrieve all update types and keep updating the channel", async () => {
     const setupState = createTestChannelState("setup");
-    await store.saveChannelState(setupState, {
-      channelFactoryAddress: setupState.networkContext.channelFactoryAddress,
-      chainId: setupState.networkContext.chainId,
-      aliceSignature: setupState.latestUpdate.aliceSignature,
-      bobSignature: setupState.latestUpdate.bobSignature,
-      state: setupState,
-    });
+    await store.saveChannelState(setupState);
 
     let fromStore = await store.getChannelState(setupState.channelAddress);
     expect(fromStore).to.deep.eq(setupState);
@@ -206,13 +172,7 @@ describe("store", () => {
       nonce: setupState.nonce + 1,
       balances: [updatedBalanceForDeposit, setupState.balances[0]],
     });
-    await store.saveChannelState(depositState, {
-      channelFactoryAddress: depositState.networkContext.channelFactoryAddress,
-      chainId: depositState.networkContext.chainId,
-      aliceSignature: depositState.latestUpdate.aliceSignature,
-      bobSignature: depositState.latestUpdate.bobSignature,
-      state: depositState,
-    });
+    await store.saveChannelState(depositState);
 
     fromStore = await store.getChannelState(setupState.channelAddress);
     expect(fromStore).to.deep.eq(depositState);
@@ -238,17 +198,7 @@ describe("store", () => {
         nonce: depositState.nonce + 1,
       },
     });
-    await store.saveChannelState(
-      createState,
-      {
-        channelFactoryAddress: createState.networkContext.channelFactoryAddress,
-        chainId: createState.networkContext.chainId,
-        aliceSignature: createState.latestUpdate.aliceSignature,
-        bobSignature: createState.latestUpdate.bobSignature,
-        state: createState,
-      },
-      transfer,
-    );
+    await store.saveChannelState(createState, transfer);
 
     fromStore = await store.getChannelState(setupState.channelAddress);
     expect(fromStore).to.deep.eq(createState);
@@ -262,13 +212,7 @@ describe("store", () => {
         },
       },
     });
-    await store.saveChannelState(resolveState, {
-      channelFactoryAddress: resolveState.networkContext.channelFactoryAddress,
-      chainId: resolveState.networkContext.chainId,
-      aliceSignature: resolveState.latestUpdate.aliceSignature,
-      bobSignature: resolveState.latestUpdate.bobSignature,
-      state: resolveState,
-    });
+    await store.saveChannelState(resolveState);
 
     fromStore = await store.getChannelState(setupState.channelAddress);
     expect(fromStore).to.deep.eq(resolveState);
@@ -298,17 +242,7 @@ describe("store", () => {
       transferTimeout: createState.latestUpdate.details.transferTimeout,
       transferState: createState.latestUpdate.details.transferInitialState,
     });
-    await store.saveChannelState(
-      createState,
-      {
-        channelFactoryAddress: createState.networkContext.channelFactoryAddress,
-        chainId: createState.networkContext.chainId,
-        aliceSignature: createState.latestUpdate.aliceSignature,
-        bobSignature: createState.latestUpdate.bobSignature,
-        state: createState,
-      },
-      transfer,
-    );
+    await store.saveChannelState(createState, transfer);
     let transferFromStore = await store.getTransferState(transfer.transferId);
     expect(transferFromStore).to.deep.eq(transfer);
 
@@ -324,18 +258,12 @@ describe("store", () => {
     resolveState.nonce = createState.nonce + 1;
     resolveState.latestUpdate.nonce = createState.latestUpdate.nonce + 1;
 
-    await store.saveChannelState(resolveState, {
-      channelFactoryAddress: resolveState.networkContext.channelFactoryAddress,
-      chainId: resolveState.networkContext.chainId,
-      aliceSignature: resolveState.latestUpdate.aliceSignature,
-      bobSignature: resolveState.latestUpdate.bobSignature,
-      state: resolveState,
-    });
+    await store.saveChannelState(resolveState);
     const fromStore = await store.getChannelState(resolveState.channelAddress);
     expect(fromStore).to.deep.eq(resolveState);
 
     transferFromStore = await store.getTransferState(transfer.transferId);
-    expect(transferFromStore.transferResolver).to.deep.eq(
+    expect(transferFromStore?.transferResolver).to.deep.eq(
       (resolveState.latestUpdate.details as ResolveUpdateDetails).transferResolver,
     );
   });
@@ -362,17 +290,7 @@ describe("store", () => {
 
     transfer1.transferResolver = undefined;
 
-    await store.saveChannelState(
-      createState,
-      {
-        channelFactoryAddress: createState.networkContext.channelFactoryAddress,
-        chainId: createState.networkContext.chainId,
-        aliceSignature: createState.latestUpdate.aliceSignature,
-        bobSignature: createState.latestUpdate.bobSignature,
-        state: createState,
-      },
-      transfer1,
-    );
+    await store.saveChannelState(createState, transfer1);
 
     const transfer2 = createTestFullHashlockTransferState({
       channelAddress: createState.channelAddress,
@@ -397,17 +315,7 @@ describe("store", () => {
       nonce: createState.latestUpdate.nonce + 1,
     });
 
-    await store.saveChannelState(
-      updatedState,
-      {
-        channelFactoryAddress: createState.networkContext.channelFactoryAddress,
-        chainId: createState.networkContext.chainId,
-        aliceSignature: createState.latestUpdate.aliceSignature,
-        bobSignature: createState.latestUpdate.bobSignature,
-        state: updatedState,
-      },
-      transfer2,
-    );
+    await store.saveChannelState(updatedState, transfer2);
 
     const channelFromStore = await store.getChannelState(createState.channelAddress);
     expect(channelFromStore).to.deep.eq(updatedState);
@@ -443,17 +351,7 @@ describe("store", () => {
 
     transfer1.transferResolver = undefined;
 
-    await store.saveChannelState(
-      createState,
-      {
-        channelFactoryAddress: createState.networkContext.channelFactoryAddress,
-        chainId: createState.networkContext.chainId,
-        aliceSignature: createState.latestUpdate.aliceSignature,
-        bobSignature: createState.latestUpdate.bobSignature,
-        state: createState,
-      },
-      transfer1,
-    );
+    await store.saveChannelState(createState, transfer1);
 
     const newBob = getRandomIdentifier();
     const transfer2 = createTestFullHashlockTransferState({
@@ -480,17 +378,7 @@ describe("store", () => {
       },
     });
 
-    await store.saveChannelState(
-      createState2,
-      {
-        channelFactoryAddress: createState2.networkContext.channelFactoryAddress,
-        chainId: createState2.networkContext.chainId,
-        aliceSignature: createState2.latestUpdate.aliceSignature,
-        bobSignature: createState2.latestUpdate.bobSignature,
-        state: createState2,
-      },
-      transfer2,
-    );
+    await store.saveChannelState(createState2, transfer2);
 
     const transfers = await store.getTransfersByRoutingId(transfer2.meta.routingId);
     expect(transfers.length).to.eq(2);