+++ modules/types/src/channel.ts
@@ -72,6 +72,11 @@ export type Balance = {
   to: Address[];
 };
 
+export enum ChannelCommitmentTypes {
+  ChannelState,
+  WithdrawData,
+}
+
 export const CoreChannelStateEncoding = tidy(`tuple(
   address channelAddress,
   address alice,
@@ -109,14 +114,6 @@ export type FullChannelState<T extends UpdateType = any> = CoreChannelState & {
   inDispute: boolean;
 };
 
-export interface ChannelCommitmentData {
-  state: CoreChannelState;
-  aliceSignature?: string;
-  bobSignature?: string;
-  channelFactoryAddress: Address;
-  chainId: number;
-}
-
 export const CoreTransferStateEncoding = tidy(`tuple(
   address channelAddress,
   bytes32 transferId,
@@ -140,12 +137,12 @@ export interface CoreTransferState {
   initialStateHash: string;
 }
 
-export type FullTransferState<T extends TransferName = any> = CoreTransferState & {
+export type FullTransferState = CoreTransferState & {
   channelFactoryAddress: string; // networkContext?
   chainId: number;
-  transferEncodings: typeof TransferEncodingsMap[T]; // Initial state encoding, resolver encoding
-  transferState: TransferStateMap[T];
-  transferResolver?: TransferResolverMap[T]; // undefined iff not resolved
+  transferEncodings: string[]; // Initial state encoding, resolver encoding
+  transferState: any;
+  transferResolver?: any; // undefined iff not resolved
   meta?: any;
   inDispute: boolean;
 };