+++ modules/documentation/docs/changelog.md
@@ -2,11 +2,20 @@
 
 ## Next Release
 
+## @connext/{types,utils,contracts,protocol,engine,browser-node}@0.0.9, @connext/{server-node}@0.0.10
+
+- introduce and use commitment types
+- reconcile hashing
+- remove commitment type
+
+## @connext/{types,utils,contracts,protocol,engine,browser-node}@0.0.9, @connext/{server-node}@0.0.10
+
 - implement generalized withdrawals
 - fix and minor improvements around asset transfers
 - move dispute methods into engine
 - shorter revert message style and contract tests speedup
 - minor cleanup and reorganization related to types and interfaces
+- make transfer types generic
 
 ## @connext/{types,utils,contracts,protocol,engine,browser-node}@0.0.9, @connext/{server-node}@0.0.9